+++ README.md
@@ -83,12 +83,14 @@ Start the server
         - 강의 생성  ✅
         - 리프레시 토큰 처리 ✅
         - 강의 종료 ✅
+        - 로그아웃 ✅
         - 실시간 코드 모니터링 💤
         - 학생 창 제어 💤
         - 일괄 코드 및 드론 실헹 제어 💤
         - 개별 코드 및 드론 실행 제어 💤
     - 학생
         - 강의 접속 💤
+        - 토큰 구현 💤
         - 코드 작성 💤
         - 코드 실행 💤
         - 드론 제어 💤
