+++ src/pages/instructor/controlMenu/CloseLectureButton.tsx
@@ -4,19 +4,15 @@ import CloseButton from "../../../shared/ui/CloseButton.tsx";
 import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import { useDeactivateLecture } from "../../../features/lecture/hooks/api/useDeactivateLecture.ts";
 import { useLecture } from "../../../shared/context/lectureProvider.tsx";
+import useCloseLecture from "../../../features/lecture/hooks/useCloseLecture.ts";
 
 const CloseLectureButton = () => {
-  const [isCloseLectureModalOpen, setIsCloseLectureModalOpen] = useState(false);
-  const { resetSavedLecture, hasSavedLecture, savedLecture } = useLecture();
-  const { mutate, data } = useDeactivateLecture();
-  const handleClickCloseButton = () => {
-    setIsCloseLectureModalOpen(false);
-    console.log("엥엥");
-    mutate({ lectureId: savedLecture.lectureId, active: false });
-    console.log("data", data);
-    resetSavedLecture();
-    /* 학생들과 통신 종료 */
-  };
+  const {
+    isCloseLectureModalOpen,
+    setIsCloseLectureModalOpen,
+    handleClickCloseButton,
+    hasSavedLecture,
+  } = useCloseLecture();
   return (
     <>
       <button
@@ -26,37 +22,36 @@ const CloseLectureButton = () => {
       >
         수업 종료
       </button>
-      {isCloseLectureModalOpen && (
-        <Modal
-          onClose={() => setIsCloseLectureModalOpen(false)}
-          isOpen={isCloseLectureModalOpen}
-        >
-          <div className="w-full h-full flex flex-col justify-center items-center space-y-4 text-2xl py-7">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 24 24"
-              strokeWidth={1.5}
-              stroke="currentColor"
-              className="size-28 text-rose-500"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126ZM12 15.75h.007v.008H12v-.008Z"
-              />
-            </svg>
 
-            <p>수업을 종료하시면 모든 학생들의 접속이 종료됩니다.</p>
-            <p>수업을 종료하시겠습니까?</p>
-            <CloseButton
-              onClick={handleClickCloseButton}
-              title={"종료하기"}
-              className={"!w-1/3 !mt-10 !mb-1.5"}
+      <Modal
+        onClose={() => setIsCloseLectureModalOpen(false)}
+        isOpen={isCloseLectureModalOpen}
+      >
+        <div className="w-full h-full flex flex-col justify-center items-center space-y-4 text-2xl py-7">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 24 24"
+            strokeWidth={1.5}
+            stroke="currentColor"
+            className="size-28 text-rose-500"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              d="M12 9v3.75m-9.303 3.376c-.866 1.5.217 3.374 1.948 3.374h14.71c1.73 0 2.813-1.874 1.948-3.374L13.949 3.378c-.866-1.5-3.032-1.5-3.898 0L2.697 16.126ZM12 15.75h.007v.008H12v-.008Z"
             />
-          </div>
-        </Modal>
-      )}
+          </svg>
+
+          <p>수업을 종료하시면 모든 학생들의 접속이 종료됩니다.</p>
+          <p>수업을 종료하시겠습니까?</p>
+          <CloseButton
+            onClick={handleClickCloseButton}
+            title={"종료하기"}
+            className={"!w-1/3 !mt-10 !mb-1.5"}
+          />
+        </div>
+      </Modal>
     </>
   );
 };
