+++ src/pages/instructor/controlMenu/LogoutButton.tsx
@@ -1,18 +1,23 @@
 import { useState } from "react";
-
 import { useLecture } from "../../../shared/context/lectureProvider.tsx";
+import { useNavigate } from "react-router-dom";
 
 const LogoutButton = () => {
   const [isControlOpen, setIsControlOpen] = useState(false);
-
   const { hasSavedLecture } = useLecture();
+  const navigate = useNavigate();
+  const handleClickLogout = () => {
+    setIsControlOpen(!isControlOpen);
+    sessionStorage.clear();
+    navigate("/", { replace: true });
+  };
 
   return (
     <>
       {!hasSavedLecture && (
         <button
           className="relative py-3 px-6 font-semibold text-lg bg-rose-500 text-white rounded-xl shadow-lg disabledBtn"
-          onClick={() => setIsControlOpen(!isControlOpen)}
+          onClick={handleClickLogout}
           disabled={hasSavedLecture}
         >
           로그아웃
